requirements.txt
pefile==2023.2.7
phonenumbers==8.13.32
Pillow==9.5.0
platformdirs==4.2.0
playsound==1.3.0
plyer==2.1.0
proglog==0.1.10
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.2
psutil==5.9.7
pure-eval==0.2.2
pvporcupine==3.0.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyAudio==0.2.14
PyAutoGUI==0.9.54
pycparser==2.21
pycryptodomex==3.20.0
pydantic==2.5.2
pydantic_core==2.14.5
pygame==2.5.2
PyGetWindow==0.0.9
Pygments==2.17.2
pyinstaller==5.13.0
pyinstaller-hooks-contrib==2023.5
pyjokes==0.6.0
PyMsgBox==1.0.9
pynput==1.7.6
pyparsing==3.1.1
PyPDF2==3.0.1
pyperclip==1.8.2
pypiwin32==223
pypng==0.20220715.0
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
PyRect==0.2.0
PyScreeze==0.1.30
PySocks==1.7.1
pytesseract==0.3.10
python-dateutil==2.8.2
pyttsx3==2.90
pytube==15.0.0
pytweening==1.0.7
pytz==2023.3
pywhatkit==5.4
pywikihow==0.5.7
pywin32==306
pywin32-ctypes==0.2.2
PyYAML==6.0.1
pyzmq==25.1.2
qrcode==7.4.2
ratelim==0.1.6
regex==2023.12.25
requests==2.28.1
requests-cache==1.2.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rfc3986==1.5.0
rsa==4.9
safetensors==0.4.2
selenium==4.16.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
speak==0.1
SpeechRecognition==3.10.1
speedtest==0.0.1
speedtest-cli==2.1.3
SQLAlchemy==2.0.25
stack-data==0.6.3
sympy==1.12
tenacity==8.2.3
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-intel==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
thread6==0.2.0
tokenizers==0.15.2
torch==2.2.1
torchvision==0.17.1
tornado==6.4
tqdm==4.66.1
traitlets==5.14.1
transformers==4.39.3
trio==0.24.0
trio-websocket==0.11.1
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
url-normalize==1.4.3
urllib3==1.26.18
wcwidth==0.2.13
Werkzeug==3.0.1
whichcraft==0.6.1
wikipedia==1.4.0
Wikipedia-API==0.6.0
wolframalpha==5.0.0
wrapt==1.14.1
wsproto==1.2.0
xmltodict==0.13.0
xyzservices==2023.10.1
yarl==1.9.4
zipp==3.18.1
zope.event==5.0
zope.interface==6.2